% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewise_residuals.R
\name{casewise_residuals}
\alias{casewise_residuals}
\title{Casewise Residuals}
\usage{
casewise_residuals(fit, add_path = FALSE)
}
\arguments{
\item{fit}{The output of a structural equation model
analysis. Only support the output of \code{\link[lavaan:sem]{lavaan::sem()}}
for now.}

\item{add_path}{Logical. If a path is not in a model
(fixed to zero), whether this path will be added to
the model #' to compute the residuals. Default is
\code{FALSE}. Use this only if necessary because the model
will be fitted once for each such path.}
}
\value{
A \code{semdplot_residuals}-class object.

TODO: Write methods for this class.
}
\description{
It computes casewise residuals
to be used by the plot functions.
}
\details{
It receives the output of structural equation
model analysis (e.g, the output of \code{\link[lavaan:sem]{lavaan::sem()}}) and
and computes casewise residuals.

It currently only supports the output of \code{\link[lavaan:sem]{lavaan::sem()}}
and a single-sample model with observable variables only.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
