% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_diag.R
\name{fit_to_diag}
\alias{fit_to_diag}
\title{Diagnostic Information}
\usage{
fit_to_diag(fit, add_path = FALSE)
}
\arguments{
\item{fit}{Argument description.}

\item{add_path}{Logical. If a path is not in a model
(fixed to zero), will this path be added to the model
to generate the diagnostic information. Default is
\code{FALSE}.}
}
\value{
A \code{semdplot}-class object.

TODO: Write methods for this class.
}
\description{
It computes diagnostic information such
residuals to be used by the plot functions.
}
\details{
It receives the output of structural equation
model analysis (e.g, the output of \code{\link[lavaan:sem]{lavaan::sem()}}) and
and computes diagnostic information to be used for the
diagnostics plots.

It currently only supports the output of \code{\link[lavaan:sem]{lavaan::sem()}}
and a single-sample model with observable variables only.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
