% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{partial_to_resid}
\alias{partial_to_resid}
\title{Residual-Only}
\usage{
partial_to_resid(diag_info)
}
\arguments{
\item{diag_info}{Argument description.}
}
\value{
A data frame
}
\description{
It extracts the residuals from
the output of \code{\link[=semdplot]{semdplot()}}.
}
\details{
It currently supports an output of any function
supported by \code{\link[=semdplot]{semdplot()}}.

TODO: Allow users to customize the plot.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
