% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag.R
\name{plot.semdplot_residuals}
\alias{plot.semdplot_residuals}
\title{Residual Plot Matrix}
\usage{
\method{plot}{semdplot_residuals}(
  x,
  what = c("partial", "residual"),
  x_names = NULL,
  y_names = NULL,
  same_x_range = TRUE,
  same_y_range = TRUE,
  partial_point_aes = list(),
  partial_loess_aes = list(),
  partial_reg_aes = list(),
  partial_b_aes = list(),
  partial_theme_aes = list(),
  added_point_aes = list(),
  added_loess_aes = list(),
  added_reg_aes = list(),
  added_b_aes = list(),
  added_theme_aes = list(),
  resid_hist_aes = list(),
  resid_density_aes = list(),
  resid_theme_aes = list(),
  text_nopath = "--",
  output_type = c("plot", "list"),
  ...
)
}
\arguments{
\item{x}{A \code{semdplot_residuals}-class object,
such as the output of \code{\link[=casewise_residuals]{casewise_residuals()}}.
Named \code{x} as required for the \code{plot} method.}

\item{what}{String. What to plot. Can be \code{"partial"}
or \code{"residual"}. if \code{"partial"}, the default,
a matrix of partial plots and/or plots of
residuals will be plotted. If \code{"residual"},
then only the plots of residuals of
y-variable will be plotted.}

\item{x_names}{Character vector. The \code{x}-variables. If
the vector is named, the names will be used instead
of the variable names in the plots.}

\item{y_names}{Character vector. The \code{y}-variables. If
the vector is named, the names will be used instead
of the variable names in the plots.}

\item{same_x_range}{Whether all plots will have the
same range for the x-axis. Default is \code{TRUE}.}

\item{same_y_range}{Whether all plots will have the
same range for the y-axis, except for the plots of
residuals. Default is \code{TRUE}.}

\item{partial_point_aes}{A list of argument for
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Used to modify the default
aesthetics of the points in the partial plot.}

\item{partial_loess_aes}{A list of argument for
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for the loess line.
Used to modify the default
aesthetics of the line in the partial plot.}

\item{partial_reg_aes}{A list of argument for
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for the regression line.
Used to modify the default
aesthetics of the line in the partial plot.}

\item{partial_b_aes}{A list of argument for
\code{\link[ggplot2:annotate]{ggplot2::annotate()}} for the standardized
regression coefficient.
Used to modify the default
aesthetics of the text in the partial plot.}

\item{partial_theme_aes}{A list of argument for
\code{\link[ggplot2:theme]{ggplot2::theme()}} for the partial plots.
Used to modify the default
theme of the partial plot.}

\item{added_point_aes}{A list of argument for
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}. Used to modify the default
aesthetics of the points in the added-variable plot.}

\item{added_loess_aes}{A list of argument for
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for the loess line.
Used to modify the default
aesthetics of the line in the added-variable plot.}

\item{added_reg_aes}{A list of argument for
\code{\link[ggplot2:geom_path]{ggplot2::geom_line()}} for the regression line.
Used to modify the default
aesthetics of the line in the added-variable plot.}

\item{added_b_aes}{A list of argument for
\code{\link[ggplot2:annotate]{ggplot2::annotate()}} for the standardized
regression coefficient.
Used to modify the default
aesthetics of the text in the added-variable plot.}

\item{added_theme_aes}{A list of argument for
\code{\link[ggplot2:theme]{ggplot2::theme()}} for the added-variable plots.
Used to modify the default
theme of the added-variable plot.}

\item{resid_hist_aes}{A list of arguments for
\code{\link[ggplot2:geom_histogram]{ggplot2::geom_histogram()}} for the plot of residuals.
Use to modify the default aesthetics of the histogram.}

\item{resid_density_aes}{A list of arguments for
\code{\link[ggplot2:geom_density]{ggplot2::geom_density()}} for the plot of residuals.
Use to modify the default aesthetics of the density plot.}

\item{resid_theme_aes}{A list of argument for
\code{\link[ggplot2:theme]{ggplot2::theme()}} for the plots of residuals.
Used to modify the default
theme of the plot of residual.}

\item{text_nopath}{String. Text to display in empty
cells. Default is`"--".}

\item{output_type}{The type of output. If \code{"plot"},
the default, the plots will be processed by
\code{\link[patchwork:wrap_plots]{patchwork::wrap_plots()}}. If \code{"list"}, the list
of plots will be returned.}

\item{...}{Additional arguments. Not used.}
}
\value{
A \code{ggplot2} plot, or a list of \code{ggplot2} plots.
}
\description{
The plot method of a
\code{semdplot_residuals}-class object.
It generates the residual plots based
on the output of \code{\link[=casewise_residuals]{casewise_residuals()}}.
}
\details{
It currently supports the output of
\code{\link[=casewise_residuals]{casewise_residuals()}}.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
