% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag.R
\name{plot_resid}
\alias{plot_resid}
\title{Residual-Only Plot Matrix}
\usage{
plot_resid(c_resid, y = NULL, progress = FALSE, ...)
}
\arguments{
\item{c_resid}{Argument description.}

\item{y}{Character. The \code{y}-variable.}

\item{progress}{To be passed to \code{\link[GGally:ggpairs]{GGally::ggpairs()}}}

\item{...}{Arguments to be passed to \code{\link[GGally:ggpairs]{GGally::ggpairs()}}.}
}
\value{
A \code{gg}-class object generated by \code{\link[GGally:ggpairs]{GGally::ggpairs()}}.
Note that \code{\link[=print]{print()}} is used to "plot" the plot.
}
\description{
It generate the plots of residuals based
on the output of \code{\link[=casewise_residuals]{casewise_residuals()}}.
}
\details{
TODO: Should be converted to a method.

It currently supports an output of any function
supported by \code{\link[=casewise_residuals]{casewise_residuals()}}.

TODO: Allow users to customize the plot.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
