% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semdplot-package.R
\docType{package}
\name{semdplot-package}
\alias{semdplot}
\alias{semdplot-package}
\title{semdplot: Diagnostic Plots for Structural Equation Models}
\description{
Diagnostic plots for structural equation models, such as plots similar to partial plots and aded-variable plots in linear regression.
}
\author{
\strong{Maintainer}: Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})

Authors:
\itemize{
  \item Mark Hok Chio Lai \email{marklhc@gmail.com } (\href{https://orcid.org/0000-0002-9196-7406}{ORCID})
  \item Rongwei Sun \email{irissun_s@hotmail.com} (\href{https://orcid.org/0000-0003-0034-1422}{ORCID})
}

}
\keyword{internal}
