% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{y_residuals}
\alias{y_residuals}
\title{Residual-Only}
\usage{
y_residuals(diag_info)
}
\arguments{
\item{diag_info}{Argument description.}
}
\value{
A data frame
}
\description{
It extracts the residuals from
the output of \code{\link[=casewise_residuals]{casewise_residuals()}}.
}
\details{
It currently supports an output of any function
supported by \code{\link[=casewise_residuals]{casewise_residuals()}}.

TODO: Allow users to customize the plot.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
