% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diag.R
\name{plot_diag}
\alias{plot_diag}
\title{Diagnostic Plot Matrix}
\usage{
plot_diag(diag_info, x = NULL, y = NULL)
}
\arguments{
\item{diag_info}{Argument description.}

\item{x}{Character. The \code{x}-variable.}

\item{y}{Character. The \code{y}-variable.}
}
\value{
A \code{ggplot2} plot, or an array of \code{ggplot2} plots.
}
\description{
It generate the diagnostic plots based
on the output of \code{\link[=fit_to_diag]{fit_to_diag()}}.
}
\details{
TODO: Should be converted to a method.

It currently output of any function supported by
\code{\link[=fit_to_diag]{fit_to_diag()}}.

TODO: Allow users to customize the plot.
}
\examples{
\donttest{
# TODO: Prepare some examples.
}

}
